+++ lib/core/data/color_xml.dart
@@ -1,0 +1,6 @@
+final class ColorXml {
+  ColorXml({required this.colors, required this.fileName});
+
+  final String fileName;
+  final List<String> colors;
+}