+++ lib/src/commands/command_args.dart
@@ -19,3 +19,19 @@ class AnalyzeCommandArgs extends CommandArgs {
     return 'AnalyzeCommandArgs instance.\nData:\nverbose - $verbose\nanalyze arb - $analyzeArb';
   }
 }
+
+class GenerateCommandArgs extends CommandArgs {
+  GenerateCommandArgs({
+    required super.verbose,
+    required this.generateColors,
+    required this.generateTextStyles,
+  });
+
+  final bool generateColors;
+  final bool generateTextStyles;
+
+  @override
+  String toString() {
+    return 'GenerateCommandArgs instance.\nData:\nverbose - $verbose\ngenerate colors - $generateColors\ngenerate text styles - $generateTextStyles';
+  }
+}