+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.4.0
+
+### Added
+
+- App Text Styles theme generation from text styles constant.
+
+## 1.3.0
+
+### Added
+
+- App colors theme generation from assets *xml.
+
 ## 1.2.1
 
 ### Added